Extract notification settings section
